Implement StubTransport for testing purposes
