+++ compute_horde/compute_horde/transport/__init__.py
@@ -1,8 +1,10 @@
 from .base import AbstractTransport, TransportConnectionError
+from .stub import StubTransport
 from .ws import WSTransport
 
 __all__ = [
     "AbstractTransport",
     "TransportConnectionError",
     "WSTransport",
+    "StubTransport",
 ]
